requirements.txt
flask
torch
torchvision
opencv-python
numpy
matplotlib
pillow
jinja2
werkzeug
click
itsdangerous